+++ client/src/App.js
@@ -6,6 +6,7 @@ import {
   createHttpLink,
 } from "@apollo/client";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
+import { setContext } from '@apollo/client/link/context';
 
 // import Auth from "./utils/auth";
 
@@ -16,17 +17,28 @@ import Login from "./components/Login";
 import NoMatch from "./components/pages/NoMatch";
 // import Developers from "../pages/Developers";
 // import Jobs from "../pages/Jobs";
-// import Profile from "../pages/Profile";
-// import Login from "../Login";
+import Profile from  "./components/pages/Profile";
 
 const httpLink = createHttpLink({
   uri: "/graphql",
 });
 
+const authLink = setContext((_, { headers }) => {
+  const token = localStorage.getItem('id_token');
+  return {
+    headers: {
+      ...headers,
+      authorization: token ? `Bearer ${token}` : '',
+    },
+  };
+});
+
 const client = new ApolloClient({
-  link: httpLink,
+  link: authLink.concat(httpLink),
   cache: new InMemoryCache(),
 });
+
+
 
 function App() {
   // once signup/login page complete
@@ -43,6 +55,10 @@ function App() {
             <Route path="/login" element={<Login />} />
             <Route path="/signup" element={<Signup />} />
             <Route path="*" element={<NoMatch />} />
+            <Route path="/profile">
+              <Route path=":username" element={<Profile />} />
+              <Route path="" element={<Profile />} />
+            </Route>
           </Routes>
         </div>
       </Router>
